+++ enterprise/__init__.py
@@ -2,4 +2,4 @@
 Your project description goes here.
 """
 
-__version__ = "4.5.1"
+__version__ = "4.6.1"