+++ enterprise/models.py
@@ -416,6 +416,12 @@ class EnterpriseCustomer(TimeStampedModel):
         help_text=_("Specifies whether the organization should have access to executive education 2U content.")
     )
 
+    enable_demo_data_for_analytics_and_lpr = models.BooleanField(
+        verbose_name="Enable demo data from analytics and lpr",
+        default=False,
+        help_text=_("Display Demo data from analyitcs and learner progress report for demo customer.")
+    )
+
     contact_email = models.EmailField(
         verbose_name="Customer admin contact email:",
         null=True,
@@ -3655,6 +3661,12 @@ class ChatGPTResponse(TimeStampedModel):
 
     .. no_pii:
     """
+    LEARNER_PROGRESS = 'learner_progress'
+    LEARNER_ENGAGEMENT = 'learner_engagement'
+    PROMPT_TYPES = [
+        (LEARNER_PROGRESS, 'Learner progress'),
+        (LEARNER_ENGAGEMENT, 'Learner engagement'),
+    ]
     uuid = models.UUIDField(primary_key=True, default=uuid4, editable=False)
 
     enterprise_customer = models.ForeignKey(
@@ -3671,6 +3683,12 @@ class ChatGPTResponse(TimeStampedModel):
     prompt = models.TextField(help_text=_('ChatGPT prompt.'))
     prompt_hash = models.CharField(max_length=32, editable=False)
     response = models.TextField(help_text=_('ChatGPT response.'))
+    prompt_type = models.CharField(choices=PROMPT_TYPES, help_text=_('Prompt type.'), max_length=32, null=True)
+
+    class Meta:
+        app_label = 'enterprise'
+        verbose_name = _('ChatGPT Response')
+        verbose_name_plural = _('ChatGPT Responses')
 
     def save(self, *args, **kwargs):
         """
@@ -3680,7 +3698,7 @@ class ChatGPTResponse(TimeStampedModel):
         super().save(*args, **kwargs)
 
     @classmethod
-    def get_or_create(cls, prompt, role, enterprise_customer):
+    def get_or_create(cls, prompt, role, enterprise_customer, prompt_type):
         """
         Get or create ChatGPT response against given prompt.
 
@@ -3691,6 +3709,7 @@ class ChatGPTResponse(TimeStampedModel):
             prompt (str): OpenAI prompt.
             role (str): ChatGPT role to assume for the prompt.
             enterprise_customer (EnterpriseCustomer): Enterprise customer UUId making the request.
+            prompt_type (str): Prompt type, e.g. learner_progress or learner_engagement etc.
 
         Returns:
             (str): Response against the given prompt.
@@ -3702,6 +3721,7 @@ class ChatGPTResponse(TimeStampedModel):
                 enterprise_customer=enterprise_customer,
                 prompt=prompt,
                 response=response,
+                prompt_type=prompt_type,
             )
             return response
         else:
@@ -4055,7 +4075,13 @@ class EnterpriseCustomerSsoConfiguration(TimeStampedModel, SoftDeletableModel):
             is_sap = True
         else:
             for field in self.base_saml_config_fields:
-                config_data[utils.camelCase(field)] = getattr(self, field)
+                if field == "active":
+                    if not updating_existing_record:
+                        config_data['enable'] = True
+                    else:
+                        config_data['enable'] = getattr(self, field)
+                else:
+                    config_data[utils.camelCase(field)] = getattr(self, field)
 
         EnterpriseSSOOrchestratorApiClient().configure_sso_orchestration_record(
             config_data=config_data,