+++ integrated_channels/moodle/client.py
@@ -13,7 +13,7 @@ from django.apps import apps
 
 from integrated_channels.exceptions import ClientError
 from integrated_channels.integrated_channel.client import IntegratedChannelApiClient
-from integrated_channels.utils import generate_formatted_log
+from integrated_channels.utils import encode_data_for_logging, generate_formatted_log
 
 LOGGER = logging.getLogger(__name__)
 
@@ -333,6 +333,17 @@ class MoodleAPIClient(IntegratedChannelApiClient):
             # The grade is exported as a decimal between [0-1]
             'grades[0][grade]': completion_data['grade'] * self.enterprise_configuration.grade_scale
         }
+
+        encoded_params = encode_data_for_logging(params)
+        LOGGER.info(generate_formatted_log(
+            self.enterprise_configuration.channel_code(),
+            self.enterprise_configuration.enterprise_customer.uuid,
+            user_id,
+            course_id,
+            'posting learner data to integrated channel '
+            f'integrated_channel_params_base64={encoded_params}'
+        ))
+
         return self._post(params)
 
     def create_content_metadata(self, serialized_data):
@@ -355,7 +366,13 @@ class MoodleAPIClient(IntegratedChannelApiClient):
         more_than_one_course = serialized_data.get('courses[1][shortname]')
         serialized_data['wsfunction'] = 'core_course_create_courses'
         try:
-            response = self._wrapped_post(serialized_data)
+            moodle_course_id = self._get_course_id(serialized_data['courses[0][idnumber]'])
+            # Course already exists but is hidden - make it visible
+            if moodle_course_id:
+                serialized_data['courses[0][visible]'] = 1
+                return self.update_content_metadata(serialized_data)
+            else:  # create a new course
+                response = self._wrapped_post(serialized_data)
         except MoodleClientError as error:
             # treat duplicate as successful, but only if its a single course
             # set chunk size settings to 1 if youre seeing a lot of these errors
@@ -397,11 +414,9 @@ class MoodleAPIClient(IntegratedChannelApiClient):
             rsp._content = bytearray('{"result": "Course not found."}', 'utf-8')  # pylint: disable=protected-access
             return rsp
         moodle_course_id = parsed_response['courses'][0]['id']
-        params = {
-            'wsfunction': 'core_course_delete_courses',
-            'courseids[]': moodle_course_id
-        }
-        response = self._wrapped_post(params)
+        serialized_data['wsfunction'] = 'core_course_update_courses'
+        serialized_data['courses[0][id]'] = moodle_course_id
+        response = self._wrapped_post(serialized_data)
         return response.status_code, response.text
 
     def create_assessment_reporting(self, user_id, payload):