+++ tests/test_integrated_channels/test_integrated_channel/test_transmitters/test_content_metadata.py
@@ -490,3 +490,34 @@ class TestContentMetadataTransmitter(unittest.TestCase):
 
         orphaned_content_record.refresh_from_db()
         assert orphaned_content_record.resolved
+
+    def test_content_data_transmission_dry_run_mode(self):
+        """
+        Test that a customer's configuration can run in dry run mode
+        """
+        # Set feature flag to true
+        self.enterprise_config.dry_run_mode_enabled = True
+
+        content_id_1 = 'course:DemoX'
+        channel_metadata_1 = {'update': True}
+        content_1 = factories.ContentMetadataItemTransmissionFactory(
+            content_id=content_id_1,
+            enterprise_customer=self.enterprise_config.enterprise_customer,
+            plugin_configuration_id=self.enterprise_config.id,
+            integrated_channel_code=self.enterprise_config.channel_code(),
+            enterprise_customer_catalog_uuid=self.enterprise_catalog.uuid,
+            channel_metadata=channel_metadata_1,
+            remote_created_at=datetime.utcnow()
+        )
+
+        create_payload = {}
+        update_payload = {}
+        delete_payload = {content_id_1: content_1}
+        self.delete_content_metadata_mock.return_value = (self.success_response_code, self.success_response_body)
+        transmitter = ContentMetadataTransmitter(self.enterprise_config)
+        transmitter.transmit(create_payload, update_payload, delete_payload)
+
+        # with dry_run_mode_enabled = True we shouldn't be able to call these methods
+        self.create_content_metadata_mock.assert_not_called()
+        self.update_content_metadata_mock.assert_not_called()
+        self.delete_content_metadata_mock.assert_not_called()